+++ src/weightinit/InitUniformRandomWeights.hpp
@@ -28,7 +28,7 @@ class InitUniformRandomWeights : public InitRandomWeights {
   protected:
    InitUniformRandomWeights();
    void initialize(char const *name, PVParams *params, Communicator const *comm);
-   virtual void randomWeights(float *patchDataStart, int patchIndex) override;
+   virtual void randomWeights(float *patchDataStart, long patchIndex) override;
 
    // Data members
   protected:
